Feature/environment-api-speedup (#60)
Massively speed up environment API on server and client side

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> Introduces a binary Protobuf response for environment data and broad
performance/safety improvements across server, DB, and web client.
> 
> - **API change (binary):** `GET /visualizer/api/environment/{tick}`
now returns `application/x-protobuf` (`EnvironmentHttpResponse`,
`CellHttpResponse`); adds timing headers and content length; removes
JSON DTOs (`EnvironmentResponseDto`, `CellWithCoordinates`).
> - **Metadata enrichment:** Adds `moleculeTypes` and `opcodes` mappings
to metadata JSON; exposes `Instruction.getAllInstructions()` and
`MoleculeTypeRegistry.getAllTypes()`.
> - **DB performance:** `RowPerChunkStrategy` caches chunk size per
schema and does O(1) chunk lookup by `first_tick`; clearer not-found
errors. `IEnvironmentDataReader` narrowed to low-level
`readChunkContaining` and reader implementation updated.
> - **Overflow safety:** Use `long` for total cell products and validate
against `Integer.MAX_VALUE` in CLI, `AnalyticsIndexer`,
`SimulationEngine`, `Environment`, `EnvironmentProperties`, and
`SimulationParameters` (methods/signatures updated).
> - **Visualizer client:** Adds protobufjs, parses binary responses,
resolves IDs via metadata mappings, defers cell map building for
tooltips, and reworks zoomed-out renderer to direct pixel buffer upload
for large-scene speedups.
> - **Tests:** Update unit/integration tests to Protobuf responses and
new DB strategy behavior.
> 
> <sup>Written by [Cursor
Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit
be864a236a9f990c4a05c9117dd640ce2ba50b0d. This will update automatically
on new commits. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

+++ src/main/java/org/evochora/datapipeline/api/memory/SimulationParameters.java
@@ -48,7 +48,7 @@ package org.evochora.datapipeline.api.memory;
  */
 public record SimulationParameters(
     int[] environmentShape,
-    int totalCells,
+    long totalCells,
     int maxOrganisms,
     int samplingInterval,
     int accumulatedDeltaInterval,
@@ -130,7 +130,7 @@ public record SimulationParameters(
      * @return SimulationParameters with calculated totalCells and default delta settings.
      */
     public static SimulationParameters of(int[] environmentShape, int maxOrganisms) {
-        int totalCells = 1;
+        long totalCells = 1L;
         for (int dim : environmentShape) {
             totalCells *= dim;
         }
@@ -157,7 +157,7 @@ public record SimulationParameters(
             int[] environmentShape, int maxOrganisms,
             int samplingInterval, int accumulatedDeltaInterval,
             int snapshotInterval, int chunkInterval, double estimatedDeltaRatio) {
-        int totalCells = 1;
+        long totalCells = 1L;
         for (int dim : environmentShape) {
             totalCells *= dim;
         }
@@ -177,7 +177,7 @@ public record SimulationParameters(
      *
      * @return Total cells (100% occupancy, no reduction).
      */
-    public int worstCaseCellsPerTick() {
+    public long worstCaseCellsPerTick() {
         return totalCells;
     }
     
@@ -298,8 +298,8 @@ public record SimulationParameters(
      * @return Estimated bytes per delta.
      */
     public long estimateBytesPerDelta() {
-        int changedCells = (int) Math.ceil(totalCells * estimatedDeltaRatio);
-        return (long) changedCells * BYTES_PER_CELL 
+        long changedCells = (long) Math.ceil(totalCells * estimatedDeltaRatio);
+        return changedCells * BYTES_PER_CELL 
              + (long) maxOrganisms * BYTES_PER_ORGANISM 
              + TICKDATA_WRAPPER_OVERHEAD;
     }
